pyproject.toml
[project]
name = "repl-nix-mentalhealthaid"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.9.0",
    "faker>=30.3.0",
    "matplotlib>=3.9.2",
    "numpy>=2.1.2",
    "openai>=1.51.2",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "scikit-learn",
    "seaborn>=0.13.2",
    "sentence-transformers>=1.5.2",
    "streamlit-extras>=0.4.7",
    "streamlit>=1.39.0",
]
